+++ spec/safe_polymorphic/version_spec.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
-RSpec.describe SafePolymorphic do
+RSpec.describe SafePolymorphic::VERSION do
   it 'has a version number' do
     expect(SafePolymorphic::VERSION).not_to be nil
-  end
-
-  it 'does something useful' do
-    expect(false).to eq(true)
   end
 end