+++ spec/spec_helper.rb
@@ -7,6 +7,9 @@ require 'database_cleaner/active_record'
 
 ActiveRecord::Base.establish_connection adapter: 'sqlite3', database: ':memory:'
 
+load "#{File.dirname(__FILE__)}/support/schema.rb"
+require "#{File.dirname(__FILE__)}/support/models.rb"
+
 SimpleCov.start do
   add_filter '/spec/'
 end