+++ spec/support/models.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative 'models/user'
+require_relative 'models/publisher'
+require_relative 'models/book'
+require_relative 'models/other_thing'