+++ spec/support/models/book.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Book < ActiveRecord::Base
+  validates :title, presence: true
+
+  belongs_to :owner, polymorphic: [Publisher, User]
+end