+++ spec/support/models/publisher.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Publisher < ActiveRecord::Base
+  validates :name, presence: true
+end