+++ spec/support/models/user.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class User < ActiveRecord::Base
+  validates :username, presence: true
+end