+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [0.1.0] - 2022-12-28
 
 - Initial release
+- Moved from [Belongs to Polymorphic](https://rubygems.org/gems/belongs_to_polymorphic)